fix warning (yes, I'm using -Werror now :)
Original commit message from CVS:
fix warning (yes, I'm using -Werror now :)

+++ gst/gstelementfactory.c
@@ -224,7 +224,7 @@ gst_elementfactory_create (GstElementFactory *factory,
   g_return_val_if_fail(factory != NULL, NULL);
 
   GST_DEBUG (GST_CAT_ELEMENTFACTORY,"creating element from factory \"%s\" with name \"%s\" and type %d\n", 
-             GST_OBJECT_NAME (factory), name, factory->type);
+             GST_OBJECT_NAME (factory), name, (gint) factory->type);
 
   if (!gst_plugin_feature_ensure_loaded (GST_PLUGIN_FEATURE (factory)))
     return NULL;
